Fixed width integer division bug

+++ PdfPixel.Fonts/Typeface/SfntPdfTypeface.cs
@@ -104,7 +104,7 @@ public sealed class SfntPdfTypeface : IPdfTypeface
     public bool IsGidExists(ushort gid) => (_cffFont != null) ? gid < _cffFont.Characters.Length : gid < _maxp.NumGlyphs;
 
     /// <inheritdoc/>
-    public float GetWidth(ushort gid) => (_font.Hmtx != null && gid < _font.Hmtx.Metrics.Count) ? _font.Hmtx.Metrics[gid].AdvanceWidth / _head.UnitsPerEm : 0f;
+    public float GetWidth(ushort gid) => (_font.Hmtx != null && gid < _font.Hmtx.Metrics.Count) ? _font.Hmtx.Metrics[gid].AdvanceWidth / (float)_head.UnitsPerEm : 0f;
 
     /// <inheritdoc/>
     public ReadOnlyMemory<byte> GetPath(ushort gid)
